Preperation for compilation steps
