--- root/templates/page.flat.pug
@@ -1,56 +1,0 @@
-doctype html
-html(lang='en')
-  head
-    style(type='text/css')!= css
-    script(type='text/javascript')!= "window.addload=function(func){if(window.addEventListener){window.addEventListener('load', func)}else{window.attachEvent('onload', func)}}"
-    script(type='text/javascript')!= svgpanzoom
-    script(type='text/javascript')!= svgjs
-    script(type='text/javascript')!= definefactions
-    script(type='text/javascript')!= definezones
-  body
-    div#container
-      div.bar
-      div#map
-        != svg
-        div#legend
-          each val,key in factions
-            div.hidden(id='llegend'+key)
-              div.legendColor(style='background-color:'+val.color+';')
-              span.legendTitle(style='color:'+val.color+';')= val.name
-      script(type='text/javascript')!= initializesvg
-      div#click
-      div#hover
-      div#blurb
-        each val,key in zones
-          - var color = factions[val.owner].color || '#fff'
-          - var bw = invertColor(color,true)
-          - var inverse = invertColor(color)
-          div.blurbContainer.hidden(id='blurb'+key)
-            div.bar
-            div.blurb(style='background-color:'+color+';')
-              span.blurbTitle(style='color:'+inverse+';') Name: 
-              span.blurbText(style='color:'+bw+';')= val.name
-              br/
-              span.blurbTitle(style='color:'+inverse+';') Owner: 
-              span.blurbText(style='color:'+bw+';')= val.owner
-              br/
-              span.blurbTitle(style='color:'+inverse+';') Information: 
-              span.blurbText(style='color:'+bw+';')= val.blurb
-              br/
-      div#long
-        each val,key in zones
-          - var color = factions[val.owner].color || '#fff'
-          - var bw = invertColor(color,true)
-          - var inverse = invertColor(color)
-          div.longContainer.hidden(id='long'+key)
-            div.bar
-            div.long(style='background-color:'+color+';')
-              span.longTitle(style='color:'+inverse+';') Name: 
-              span.longText(style='color:'+bw+';')= val.name
-              br/
-              span.longTitle(style='color:'+inverse+';') Owner: 
-              span.longText(style='color:'+bw+';')= val.owner
-              br/
-              span.longTitle(style='color:'+inverse+';') Information: 
-              span.longText(style='color:'+bw+';')= val.blurb
-              br/
