+++ src/javascript/define_zones.js
@@ -1,0 +1,428 @@
+var ZoneDetails = {
+  "ClawZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Claw Country"
+  },
+  "ClawZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Claw Country"
+  },
+  "ClawZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Claw Country"
+  },
+  "FangZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fang Country"
+  },
+  "KiriRuinsZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Kirigakure Ruins"
+  },
+  "IslandZone01":{
+     "name":"New Kirigakure",
+     "blurb":"",
+     "detail":"",
+     "owner":"Island Country"
+  },
+  "IslandZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Island Country"
+  },
+  "VolcanoZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Volcano Country"
+  },
+  "LightningZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Lightning Country"
+  },
+  "LightningZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Lightning Country"
+  },
+  "LightningZone04":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Lightning Country"
+  },
+  "LightningZone05":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Lightning Country"
+  },
+  "LightningZone06":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Lightning Country"
+  },
+  "BearZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Bear Country"
+  },
+  "RainZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Rain Country"
+  },
+  "RainZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Rain Country"
+  },
+  "RainZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Rain Country"
+  },
+  "FireZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone04":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone05":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone06":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone07":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "FireZone08":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Fire Country"
+  },
+  "TideZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Whirling Tide Country"
+  },
+  "FlowerZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Flower Country"
+  },
+  "TeaZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Tea Country"
+  },
+  "TeaZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Tea Country"
+  },
+  "SnowZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Snow Country"
+  },
+  "SnowZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Snow Country"
+  },
+  "SnowZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Snow Country"
+  },
+  "SnowZone04":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Snow Country"
+  },
+  "SnowZone05":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Snow Country"
+  },
+  "MoonZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Moon Country"
+  },
+  "WaterfallZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Waterfall Country"
+  },
+  "WaterfallZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Waterfall Country"
+  },
+  "GrassZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Grass Country"
+  },
+  "EarthZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "EarthZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "EarthZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "EarthZone04":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "EarthZone05":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "EarthZone06":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "EarthZone07":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Earth Country"
+  },
+  "ForestZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Forest Country"
+  },
+  "ForestZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Forest Country"
+  },
+  "ShadowZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Shadow Country"
+  },
+  "SwampZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Swamp Country"
+  },
+  "TornadoZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Tornado Country"
+  },
+  "BirdZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Bird Country"
+  },
+  "StoneZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Stone Country"
+  },
+  "IronZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Iron Country"
+  },
+  "RiverZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"River Country"
+  },
+  "RiverZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"River Country"
+  },
+  "WindZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone04":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone05":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone06":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone07":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "WindZone08":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Wind Country"
+  },
+  "HavenZone01":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  },
+  "HavenZone02":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  },
+  "HavenZone03":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  },
+  "HavenZone04":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  },
+  "HavenZone05":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  },
+  "HavenZone06":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  },
+  "HavenZone07":{
+     "name":"",
+     "blurb":"",
+     "detail":"",
+     "owner":"Haven Country"
+  }
+}
+
+if (typeof window !== 'undefined') {
+  window.ZoneDetails = ZoneDetails
+} else {
+  module.exports = ZoneDetails
+}
